Eliminated compilation warnings under Visual C++ 2005.
git-svn-id: https://tigervnc.svn.sourceforge.net/svnroot/tigervnc@2265 3789f03b-4d11-0410-bbf8-ca57d06f2519

+++ orig/branches/BRANCH_1_3_STABLE/vnc_winsrc/vncviewer/FileTransfer.cpp
@@ -588,7 +588,7 @@ FileTransfer::FileTransferUpload()
 					SendFileUploadDataMessage(mTime);
 					break;
 				}
-				SendFileUploadDataMessage(dwNumberOfBytesRead, pBuff);
+				SendFileUploadDataMessage((unsigned short)dwNumberOfBytesRead, pBuff);
 				dwPortionRead += dwNumberOfBytesRead;
 				if (dwPortionRead >= (10 * sz_rfbBlockSize)) {
 					dwPortionRead = 0;
@@ -1322,5 +1322,5 @@ void FileTransfer::Time70ToFiletime(unsigned int time70, FILETIME *pftime)
 {
     LONGLONG ll = Int32x32To64(time70, 10000000) + 116444736000000000;
     pftime->dwLowDateTime = (DWORD) ll;
-    pftime->dwHighDateTime = ll >> 32;
+    pftime->dwHighDateTime = (DWORD)(ll >> 32);
 }
